+++ grpc/server/src/plugins/action/action_service_impl.cpp
@@ -1,4 +1,4 @@
-#include "actionrpc_impl.h"
+#include "action_service_impl.h"
 
 namespace rpc_action = dronecore::rpc::action;
